[feat: #12] 가게 등록 뷰 구현
