+++ src/main/java/org/prgrms/kyu/controller/UserController.java
@@ -1,10 +1,7 @@
 package org.prgrms.kyu.controller;
 
-import javax.naming.AuthenticationException;
 import lombok.RequiredArgsConstructor;
 import org.prgrms.kyu.dto.JoinRequest;
-import org.prgrms.kyu.dto.UserInfo;
-import org.prgrms.kyu.entity.User;
 import org.prgrms.kyu.entity.UserType;
 import org.prgrms.kyu.service.SecurityService;
 import org.prgrms.kyu.service.StoreService;
@@ -53,22 +50,6 @@ public class UserController {
     public String login(Model model, String logout) {
         if (securityService.isAuthenticated()) return "redirect:/";
         if (logout != null) model.addAttribute("message", "안전하게 로그아웃되었습니다.");
-        return "/user/loginForm";
-    }
-
-    @GetMapping("/user/myStore")
-    public String myStore(Model model, Authentication authentication) {
-        if (!securityService.isAuthenticated()) return "redirect:/";
-        UserType userType = userService.getUserType(
-            ((UserDetails) authentication.getPrincipal()).getUsername());
-        if(userType.equals(UserType.STORE_OWNER)){
-            model.addAttribute("userInfo",
-                userService.getUser(((UserDetails) authentication.getPrincipal()).getUsername()));
-
-            return "/store/myStore";
-        }else if(userType.equals(UserType.CUSTOMER)){
-            return "/index";
-        }
         return "/user/loginForm";
     }
