+++ src/main/resources/templates/store/my-store.html
@@ -1,0 +1,39 @@
+<!DOCTYPE html>
+<html lang="ko" xmlns:th="http://www.thymeleaf.org">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <meta name="author" content="Mark Otto, Jacob Thornton, and Bootstrap contributors">
+    <!-- Bootstrap core CSS -->
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
+    <link href="/css/signin.css" rel="stylesheet">
+    <title>뀨의 민족 가게 등록</title>
+</head>
+<body class="text-center">
+<main class="form-signin">
+    <form method="POST" class="form-signin" th:object="${storeForm}" th:action="@{/store}">
+        <h2 class="form-store-heading">Create Store</h2><br>
+        <div class="form-group">
+            <input type="text" th:field="*{name}" class="form-control" placeholder="Store Name" autofocus="true">
+        </div>
+
+        <div class="form-group">
+            <input type="text" th:field="*{telephone}" class="form-control" placeholder="Telephone" autofocus="true">
+        </div>
+
+        <div class="form-group">
+            <input type="text" th:field="*{description}" class="form-control" placeholder="Description" autofocus="true">
+        </div>
+
+        <div class="form-group">
+            <input type="text" th:field="*{location}" class="form-control" placeholder="Location" autofocus="true">
+        </div>
+
+        <input type="hidden" id="userId" class="form-control" name="userId" th:value="${userInfo.getId()}"/>
+
+        <button class="btn btn-lg btn-primary btn-block" type="submit">가게 생성</button>
+    </form>
+</main>
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p" crossorigin="anonymous"></script>
+</body>
+</html>
